Fixing some of the test form the "tests" module

+++ core/src/test/java/com/jetbrains/youtrack/db/internal/core/metadata/schema/TestSchemaPropertyTypeDetection.java
@@ -79,7 +79,7 @@ public class TestSchemaPropertyTypeDetection extends DbTestBase {
     assertEquals(PropertyTypeInternal.DECIMAL,
         PropertyTypeInternal.getTypeByClass(BigDecimal.class));
 
-    assertEquals(PropertyTypeInternal.INTEGER,
+    assertEquals(PropertyTypeInternal.DECIMAL,
         PropertyTypeInternal.getTypeByClass(BigInteger.class));
 
     assertEquals(PropertyTypeInternal.LINK,
@@ -168,7 +168,7 @@ public class TestSchemaPropertyTypeDetection extends DbTestBase {
     assertEquals(PropertyTypeInternal.DECIMAL,
         PropertyTypeInternal.getTypeByValue(new BigDecimal(10)));
 
-    assertEquals(PropertyTypeInternal.INTEGER,
+    assertEquals(PropertyTypeInternal.DECIMAL,
         PropertyTypeInternal.getTypeByValue(new BigInteger("20")));
 
     assertEquals(PropertyTypeInternal.LINK,
